refactor: restructure subcommands
