+++ domain/src/Wolfe.SpaceTraders.Domain/Wolfe.SpaceTraders.Domain.csproj
@@ -7,12 +7,14 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Humanizer" Version="2.14.1" />
-    <PackageReference Include="StronglyTypedId" Version="1.0.0-beta06" />
+    <Compile Remove="Contracts\Commands\**" />
+    <EmbeddedResource Remove="Contracts\Commands\**" />
+    <None Remove="Contracts\Commands\**" />
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="Contracts\Commands\" />
+    <PackageReference Include="Humanizer" Version="2.14.1" />
+    <PackageReference Include="StronglyTypedId" Version="1.0.0-beta06" />
   </ItemGroup>
 
 </Project>
